+++ NeoForge/src/main/java/fuzs/puzzleslib/neoforge/api/client/data/v2/AbstractAtlasProvider.java
@@ -3,11 +3,13 @@ package fuzs.puzzleslib.neoforge.api.client.data.v2;
 import fuzs.puzzleslib.api.data.v2.core.DataProviderContext;
 import net.minecraft.client.renderer.texture.atlas.sources.SingleFile;
 import net.minecraft.client.resources.model.Material;
+import net.minecraft.client.resources.model.ModelManager;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
 import net.minecraft.resources.ResourceLocation;
 import net.neoforged.neoforge.common.data.SpriteSourceProvider;
 
+import java.util.Objects;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 
@@ -29,9 +31,8 @@ public abstract class AbstractAtlasProvider extends SpriteSourceProvider {
     public abstract void addAtlases();
 
     public void addMaterial(Material material) {
-        ResourceLocation resourceLocation = material.atlasLocation().withPath((String s) -> {
-            return s.replace("textures/atlas/", "").replace(".png", "");
-        });
+        ResourceLocation resourceLocation = ModelManager.VANILLA_ATLASES.get(material.atlasLocation());
+        Objects.requireNonNull(resourceLocation, "atlas for " + material.atlasLocation() + " is null");
         this.atlas(resourceLocation).addSource(new SingleFile(material.texture(), Optional.empty()));
     }