+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.4.2-1.21.4] - 2025-02-04
+### Added
+- Add `BlockStateResolverContext`
+- Add `ClientStartedCallback`
+- Add `ModelLoadingHelper`
+### Changed
+- Delay running `AdditionalModelsContext` until `ModelLoadingPlugin` initialization on Fabric
+### Fixed
+- Fix `AbstractModelProvider` generating all vanilla item models on each run
+
 ## [v21.4.1-1.21.4] - 2025-02-03
 ### Changed
 - Temporarily allow data generation to run by hijacking the client configuration while Architectury Loom remains broken